refactor: continue splitting tests up by required dependency.
Health checks are up.
